Move instructions into separate folder and files
