+++ anchor/programs/solanavotingdapp/src/instructions/mod.rs
@@ -1,0 +1,2 @@
+pub mod initialize_poll;
+pub mod vote;
